refactor: study-room에 있는 repository와 entity 폴더 분리
