+++ server/src/study-room/entity/study-room-participant.entity.ts
