+++ server/src/study-room/entity/study-room.entity.ts
